Copy dalponte2016() GA for watershed, li2012, and silva2016

+++ GA_silva2016/Individual.h
@@ -18,7 +18,7 @@ public:
 	virtual ~Individual();
 
 	int chromLength;
-	int chromosome[MAX_CHROMLENGTH];
+	double chromosome[MAX_CHROMLENGTH];
 	double fitness;
 
 	void Init();
